requirements.txt
openai
chromadb
streamlit
fastapi
uvicorn
python-multipart
python-dotenv